Move syslog into Unix-only source code
Windows does not have syslog, so we need to use build tags to fix the
build.

Fixes dedis/cothority#1326.
